requirements.txt
arcade==3.0.0.dev27
attrs==23.2.0
beautifulsoup4==4.12.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.7
pillow==10.2.0
platformdirs==4.2.2
pycparser==2.22
pyglet==2.1.dev2
pymunk==6.6.0
python-slugify==8.0.4
pytiled-parser==2.2.3
requests==2.31.0
requests-cache==1.2.0
six==1.16.0
soupsieve==2.5
text-unidecode==1.3
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==2.2.1
